requirements.txt
# Core dependencies
numpy>=1.21.0
tensorflow>=2.13.0
keras>=2.13.0
seaborn>=0.11.0

# Computer Vision
opencv-contrib-python>=4.8.0
Pillow>=9.0.0

# PyTorch ecosystem
torch>=1.9.0
torchvision>=0.10.0
facenet-pytorch>=2.0.1

# Machine Learning
scikit-learn>=1.0.0
scikit-image>=0.19.0

# Data handling
pandas>=1.3.0
imageio>=2.15.0

# Visualization
matplotlib>=3.5.0

# Keras add-ons
keras_efficientnets>=0.1.6

# Optional (if using specific features)
keras-facenet==0.1a5
keras-vis==0.4.1